backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pdfplumber==0.10.3
python-docx==0.8.11
ollama==0.1.7
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
watchfiles==0.21.0